Added mealMock.js
